+++ src/fping.c
@@ -270,7 +270,7 @@ IP_HEADER_RESULT default_ip_header_result() {
     res.otime_ms = 0x80000000U;
     res.rtime_ms = 0x80000000U;
     res.ttime_ms = 0x80000000U;
-    res.src_addr[0] = '\0';
+    res.reply_dst_addr[0] = '\0';
     return res;
 }
 
@@ -541,7 +541,7 @@ int main(int argc, char **argv)
         { "check-source", 0, OPTPARSE_NONE },
         { "print-tos", 0, OPTPARSE_NONE },
         { "print-ttl", 0, OPTPARSE_NONE },
-        { "print-srcaddr", 0, OPTPARSE_NONE },
+        { "print-reply-dst", 0, OPTPARSE_NONE },
         { "seqmap-timeout", 0, OPTPARSE_REQUIRED },
 #if defined(DEBUG) || defined(_DEBUG)
         { NULL, 'z', OPTPARSE_REQUIRED },
@@ -605,8 +605,8 @@ int main(int argc, char **argv)
                     }
                 }
 #endif
-            } else if (strstr(optparse_state.optlongname, "print-srcaddr") != NULL) {
-                opt_print_srcaddr_on = 1;
+            } else if (strstr(optparse_state.optlongname, "print-reply-dst") != NULL) {
+                opt_print_reply_dst_on = 1;
             } else if (strstr(optparse_state.optlongname, "seqmap-timeout") != NULL) {
                 opt_seqmap_timeout = strtod_strict(optparse_state.optarg) * 1000000;
             } else if (strstr(optparse_state.optlongname, "oiface") != NULL) {
@@ -2366,10 +2366,10 @@ int decode_icmp_ipv4(
         ip_header_res->ttime_ms = ntohl(icp->icmp_dun.id_ts.its_ttime);
     }
 
-    if (opt_print_srcaddr_on) {
-        if (ip == NULL || inet_ntop(AF_INET, &ip->ip_dst, ip_header_res->src_addr, sizeof(ip_header_res->src_addr)) == NULL) {
-            strncpy(ip_header_res->src_addr, "unknown", sizeof(ip_header_res->src_addr) - 1);
-            ip_header_res->src_addr[sizeof(ip_header_res->src_addr) - 1] = '\0';
+    if (opt_print_reply_dst_on) {
+        if (ip == NULL || inet_ntop(AF_INET, &ip->ip_dst, ip_header_res->reply_dst_addr, sizeof(ip_header_res->reply_dst_addr)) == NULL) {
+            strncpy(ip_header_res->reply_dst_addr, "unknown", sizeof(ip_header_res->reply_dst_addr) - 1);
+            ip_header_res->reply_dst_addr[sizeof(ip_header_res->reply_dst_addr) - 1] = '\0';
         }
     }
 
@@ -2493,9 +2493,9 @@ int decode_icmp_ipv6(
     *id = icp->icmp6_id;
     *seq = ntohs(icp->icmp6_seq);
 
-    if (opt_print_srcaddr_on) {
-        strncpy(ip_header_res->src_addr, "not supported", sizeof(ip_header_res->src_addr) - 1);
-        ip_header_res->src_addr[sizeof(ip_header_res->src_addr) - 1] = '\0';
+    if (opt_print_reply_dst_on) {
+        strncpy(ip_header_res->reply_dst_addr, "not supported", sizeof(ip_header_res->reply_dst_addr) - 1);
+        ip_header_res->reply_dst_addr[sizeof(ip_header_res->reply_dst_addr) - 1] = '\0';
     }
 
     return 1;
@@ -3165,6 +3165,6 @@ void usage(int is_error)
     fprintf(out, "   -X, --fast-reachable=N exits true immediately when N hosts are found\n");
     fprintf(out, "       --print-tos    show received TOS value\n");
     fprintf(out, "       --print-ttl    show IP TTL value\n");
-    fprintf(out, "       --print-srcaddr show used IP source address (IPv6 is currently not supported).\n");
+    fprintf(out, "       --print-reply-dst show the destination address of the received reply packet (IPv6 is currently not supported).\n");
     exit(is_error);
 }